feat: 加 /agenda 路由 + Header 议程导航
议程不再只在密码门可见。抽出 Agenda 组件供 / 和 /agenda 共用，
Header 加 "议程" tab，登录后任意页面一键到议程。
